Create component folder and relocate createCard functions inside
