+++ public/index.html
@@ -7,7 +7,7 @@
   <meta name="description" content="Find the best information of Game of Thrones characters">
   <link rel="shortcut icon" href="favicon.svg" type="image/svg+xml">
   <link rel="stylesheet" href="styles/styles.css">
-  <script type="module" src="js/index.js"></script>
+  <script type="module" src="/js/index.js"></script>
   <title>Game of Thrones</title>
 </head>
