+++ src/characters/components/card/createMainContainer.ts
