+++ src/components/App/createApp.ts
@@ -1,14 +1,14 @@
-import createMainContainer from "../../characters/character/card/createMainContainer.js";
+import { createCardContainer } from "../../characters/components/card/index.js";
 import createMainHeader from "../header/createMainHeader.js";
 
 const createApp = (): HTMLElement => {
   const appContainer = document.createElement("div");
 
   const mainHeader = createMainHeader();
-  const mainContainer = createMainContainer();
+  const cardContainer = createCardContainer();
 
   appContainer.appendChild(mainHeader);
-  appContainer.appendChild(mainContainer);
+  appContainer.appendChild(cardContainer);
 
   return appContainer;
 };
